Starting a MIR control flow analyzer
